+++ src/main/java/net/brnbrd/delightful/data/gen/DelightfulBlockTagProvider.java
@@ -47,6 +47,7 @@ public class DelightfulBlockTagProvider extends BlockTagsProvider {
 				.add(DelightfulBlocks.GLOOMGOURD_PIE.get())
 				.add(DelightfulBlocks.CHORUS_PIE.get())
 				.add(DelightfulBlocks.MULBERRY_PIE.get())
+				.add(DelightfulBlocks.PASSION_FRUIT_TART.get())
 				.add(DelightfulBlocks.BAKLAVA.get())
 				.addOptional(Modid.UG.rl("gloomgourd"))
 				.addOptional(Modid.UG.rl("carved_gloomgourd"));